Conflicts:
	app/models/table.rb

+++ app/models/user.rb
@@ -1,13 +1,14 @@
 # coding: UTF-8
+require 'cartodb/per_request_sequel_cache'
 require_relative './user/user_decorator'
 require_relative './user/oauths'
 require_relative './synchronization/synchronization_oauth'
 require_relative './visualization/member'
-require_relative './visualization/external_source'
 require_relative './visualization/collection'
 require_relative './user/user_organization'
 require_relative './synchronization/collection.rb'
-require_relative '../../app/models/common_data'
+require_relative '../services/visualization/common_data_service'
+require_relative './external_data_import'
 require_relative './feature_flag'
 
 class User < Sequel::Model
@@ -34,7 +35,7 @@ class User < Sequel::Model
   one_to_many :search_tweets, order: :created_at.desc
   many_to_one :organization
 
-  many_to_many :layers, :order => :order, :after_add => proc { |user, layer|
+  many_to_many :layers, class: ::Layer, :order => :order, :after_add => proc { |user, layer|
     layer.set_default_order(user)
   }
 
@@ -45,6 +46,7 @@ class User < Sequel::Model
   plugin :validation_helpers
   plugin :json_serializer
   plugin :dirty
+  plugin :caching, PerRequestSequelCache
 
   # Restrict to_json attributes
   @json_serializer_opts = {
@@ -137,52 +139,25 @@ class User < Sequel::Model
     setup_user
     save_metadata
     self.load_avatar
-    load_common_data if FeatureFlag.allowed?('load_common_data')
     monitor_user_notification
     sleep 1
     set_statement_timeouts
-    if self.has_organization_enabled?
-      ::Resque.enqueue(::Resque::UserJobs::Mail::NewOrganizationUser, self.id)
-    end
   end
 
-  def load_common_data(datasets = CommonDataSingleton.instance.datasets[:datasets])
-    Rollbar.report_message('common data', 'debug', {
-      :action => 'load',
-      :user_id => self.id,
-      :username => self.username
-    })
-    datasets.each do |d|
-      v = CartoDB::Visualization::Member.remote_member(
-        d['name'],
-        self.id,
-        CartoDB::Visualization::Member::PRIVACY_PUBLIC,
-        d['description'],
-        [ d['category'] ],
-        d['license'],
-        d['source'])
-      v.store
+  def notify_new_organization_user
+    ::Resque.enqueue(::Resque::UserJobs::Mail::NewOrganizationUser, self.id)
+  end
 
-      CartoDB::Visualization::ExternalSource.new(v.id, d['url'], d['geometry_types'], d['rows'], d['size'], 'common-data').save
-    end
-  rescue => e
-    Rollbar.report_exception(e)
+  def should_load_common_data?
+    last_common_data_update_date.nil? || last_common_data_update_date < Time.now - 1.month
+  end
+
+  def load_common_data
+    CartoDB::Visualization::CommonDataService.new.load_common_data_for_user(self)
   end
 
   def delete_common_data
-    CartoDB::Visualization::Collection.new.fetch({type: 'remote', user_id: self.id}).map do |v|
-      begin
-        CartoDB::Visualization::ExternalSource.where(visualization_id: v.id).delete
-        v.delete
-      rescue Sequel::DatabaseError => e
-        match = e.message =~ /violates foreign key constraint "external_data_imports_external_source_id_fkey"/
-        if match.present? && match >= 0
-          puts "Couldn't delete #{v.id} visualization because it's been imported"
-        else
-          raise e
-        end
-      end
-    end
+    CartoDB::Visualization::CommonDataService.new.delete_common_data_for_user(self)
   end
 
   def after_save
@@ -204,6 +179,19 @@ class User < Sequel::Model
 
   end
 
+  def can_delete
+    !has_shared_entities?
+  end
+
+  def has_shared_entities?
+    # Right now, cannot delete users with entities shared with other users or the org.
+    has_shared_entities = false
+    CartoDB::Permission.where(owner_id: self.id).each { |permission|
+      has_shared_entities = has_shared_entities || !permission.acl.empty?
+    }
+    has_shared_entities
+  end
+
   def before_destroy
     org_id = nil
     @org_id_for_org_wipe = nil
@@ -220,11 +208,6 @@ class User < Sequel::Model
         end
       end
 
-      # Right now, cannot delete users with entities shared with other users or the org.
-      can_delete = true
-      CartoDB::Permission.where(owner_id: self.id).each { |permission|
-        can_delete = can_delete && permission.acl.empty?
-      }
       unless can_delete
         raise CartoDB::BaseCartoDBError.new('Cannot delete user, has shared entities')
       end
@@ -240,12 +223,21 @@ class User < Sequel::Model
       self.tables.all.each { |t| t.destroy }
 
       # Remove user data imports, maps, layers and assets
+      self.delete_external_data_imports
+      self.delete_external_sources
+      # There's a FK from geocodings to data_import.id so must be deleted in proper order
+      if self.organization.nil? || self.organization.owner.nil? || self.id == self.organization.owner.id
+        self.geocodings.each { |g| g.destroy }
+      else
+        assign_geocodings_to_organization_owner
+      end
       self.data_imports.each { |d| d.destroy }
       self.maps.each { |m| m.destroy }
       self.layers.each { |l| remove_layer l }
-      self.geocodings.each { |g| g.destroy }
       self.assets.each { |a| a.destroy }
       CartoDB::Synchronization::Collection.new.fetch(user_id: self.id).destroy
+
+      assign_search_tweets_to_organization_owner
     rescue StandardError => exception
       error_happened = true
       CartoDB::Logger.info "Error destroying user #{username}. #{exception.message}\n#{exception.backtrace}"
@@ -264,11 +256,29 @@ class User < Sequel::Model
       if User.where(:database_name => self.database_name).count > 1
         raise CartoDB::BaseCartoDBError.new('The user is not supposed to be in a organization but another user has the same database_name. Not dropping it')
       else
-        drop_database_and_user unless error_happened
+        if !error_happened
+          Thread.new do
+            drop_database_and_user
+          end.join
+          monitor_user_notification
+        end
       end
     end
 
     self.feature_flags_user.each { |ffu| ffu.delete }
+  end
+
+  def delete_external_data_imports
+    external_data_imports = ExternalDataImport.by_user_id(self.id)
+    external_data_imports.each { |edi| edi.destroy }
+  rescue => e
+    Rollbar.report_message('Error deleting external data imports at user deletion', 'error', { user: self.inspect, error: e.inspect })
+  end
+
+  def delete_external_sources
+    delete_common_data
+  rescue => e
+    Rollbar.report_message('Error deleting external data imports at user deletion', 'error', { user: self.inspect, error: e.inspect })
   end
 
   def after_destroy
@@ -278,7 +288,7 @@ class User < Sequel::Model
     end
   end
 
-  # Org users share the same db, so must only delete the schema
+  # Org users share the same db, so must only delete the schema unless he's the owner
   def drop_organization_user(org_id, is_owner=false)
     raise CartoDB::BaseCartoDBError.new('Tried to delete an organization user without org id') if org_id.nil?
 
@@ -315,21 +325,23 @@ class User < Sequel::Model
       conn.run("REVOKE ALL ON DATABASE \"#{database_name}\" FROM \"#{database_username}\"")
       conn.run("REVOKE ALL ON DATABASE \"#{database_name}\" FROM \"#{database_public_username}\"")
       conn.run("DROP USER \"#{database_public_username}\"")
-      conn.run("DROP USER \"#{database_username}\"")
+      if is_owner
+        drop_database_and_user if is_owner
+      else
+        conn.run("DROP USER \"#{database_username}\"")
+      end
+
     end.join
 
     monitor_user_notification
   end
 
   def drop_database_and_user
-    Thread.new do
-      conn = self.in_database(as: :cluster_admin)
-      conn.run("UPDATE pg_database SET datallowconn = 'false' WHERE datname = '#{database_name}'")
-      User.terminate_database_connections(database_name, database_host)
-      conn.run("DROP DATABASE \"#{database_name}\"")
-      conn.run("DROP USER \"#{database_username}\"")
-    end.join
-    monitor_user_notification
+    conn = self.in_database(as: :cluster_admin)
+    conn.run("UPDATE pg_database SET datallowconn = 'false' WHERE datname = '#{database_name}'")
+    User.terminate_database_connections(database_name, database_host)
+    conn.run("DROP DATABASE \"#{database_name}\"")
+    conn.run("DROP USER \"#{database_username}\"")
   end
 
   def self.terminate_database_connections(database_name, database_host)
@@ -402,12 +414,23 @@ class User < Sequel::Model
     @new_password = new_password_value
     @new_password_confirmation = new_password_confirmation_value
 
-    @old_password_validated = self.class.password_digest(old_password, self.salt) == self.crypted_password
+    @old_password_validated = validate_old_password(old_password)
     return unless @old_password_validated
 
     return unless new_password_value == new_password_confirmation_value && !new_password_value.nil?
 
+    # Must be set AFTER validations
+    set_last_password_change_date
+
     self.password = new_password_value
+  end
+
+  def validate_old_password(old_password)
+    (self.class.password_digest(old_password, self.salt) == self.crypted_password) || (google_sign_in && last_password_change_date.nil?)
+  end
+
+  def should_display_old_password?
+    google_sign_in.nil? || !google_sign_in || !last_password_change_date.nil?
   end
 
   def password_confirmation
@@ -415,7 +438,7 @@ class User < Sequel::Model
   end
 
   def password_confirmation=(password_confirmation)
-    self.last_password_change_date = Time.zone.now unless new?
+    set_last_password_change_date
     @password_confirmation = password_confirmation
   end
 
@@ -485,7 +508,7 @@ class User < Sequel::Model
   end
 
   def database_public_username
-    has_organization_enabled? ? "cartodb_publicuser_#{id}" : CartoDB::PUBLIC_DB_USER
+    (self.database_schema != CartoDB::DEFAULT_DB_SCHEMA) ? "cartodb_publicuser_#{id}" : CartoDB::PUBLIC_DB_USER
   end
 
   def database_password
@@ -638,7 +661,7 @@ class User < Sequel::Model
   end #map_tags
 
   def tables
-    ::Table.filter(:user_id => self.id).order(:id).reverse
+    ::UserTable.filter(:user_id => self.id).order(:id).reverse
   end
 
   def tables_including_shared
@@ -1043,25 +1066,33 @@ class User < Sequel::Model
   # TODO: Without a full table scan, ignoring the_geom_webmercator, we cannot accuratly asses table size
   # Needs to go on a background job.
   def db_size_in_bytes
+    return 0 if self.new?
+
     attempts = 0
     begin
       # Hack to support users without the new MU functiones loaded
       user_data_size_function = self.cartodb_extension_version_pre_mu? ? "CDB_UserDataSize()" : "CDB_UserDataSize('#{self.database_schema}')"
       result = in_database(:as => :superuser).fetch("SELECT cartodb.#{user_data_size_function}").first[:cdb_userdatasize]
       result
-    rescue
+    rescue => e
       attempts += 1
-      in_database(:as => :superuser).fetch("ANALYZE")
+      begin
+        in_database(:as => :superuser).fetch("ANALYZE")
+      rescue => ee
+        Rollbar.report_exception(ee)
+        raise ee
+      end
       retry unless attempts > 1
+      Rollbar.report_exception(e)
     end
   end
 
-  def real_tables
+  def real_tables(in_schema=self.database_schema)
     self.in_database(:as => :superuser)
     .select(:pg_class__oid, :pg_class__relname)
     .from(:pg_class)
     .join_table(:inner, :pg_namespace, :oid => :relnamespace)
-    .where(:relkind => 'r', :nspname => self.database_schema)
+    .where(:relkind => 'r', :nspname => in_schema)
     .exclude(:relname => ::Table::SYSTEM_TABLE_NAMES)
     .all
   end
@@ -1072,7 +1103,7 @@ class User < Sequel::Model
 
   def link_ghost_tables_working
     # search in the first 100. This is random number
-    enqeued = Resque.peek(:users, 0, 100).select { |job| 
+    enqeued = Resque.peek(:users, 0, 100).select { |job|
       job && job['class'] === 'Resque::UserJobs::SyncTables::LinkGhostTables' && !job['args'].nil? && job['args'].length == 1 && job['args'][0] === self.id
     }.length
     workers = Resque::Worker.all
@@ -1143,7 +1174,7 @@ class User < Sequel::Model
     metadata_table_names = self.tables.select(:name).map(&:name)
     renamed_tables       = real_tables.reject{|t| metadata_table_names.include?(t[:relname])}.select{|t| metadata_tables_ids.include?(t[:oid])}
     renamed_tables.each do |t|
-      table = ::Table.find(:table_id => t[:oid])
+      table = Table.new(:user_table => ::UserTable.find(:table_id => t[:oid]))
       begin
         Rollbar.report_message('ghost tables', 'debug', {
           :action => 'rename',
@@ -1190,20 +1221,22 @@ class User < Sequel::Model
     metadata_tables_ids = metadata_tables.select{ |table| !syncs.include?(table[:name]) }
                                          .map{ |table| table[:table_id] }
 
-    dropped_tables = metadata_tables_ids - real_tables.map{|t| t[:oid]}
+    dropped_tables = metadata_tables_ids - real_tables.map{|t| t[:oid]} - [nil]
 
     # Remove tables with oids that don't exist on the db
-    self.tables.where(table_id: dropped_tables).all.each do |table|
+    self.tables.where(table_id: dropped_tables).all.each do |user_table|
       Rollbar.report_message('ghost tables', 'debug', {
         :action => 'dropping table',
-        :new_table => table.name
+        :new_table => user_table.name
       })
+      table = Table.new(user_table: user_table)
       table.keep_user_database_table = true
       table.destroy
     end if dropped_tables.present?
 
     # Remove tables with null oids unless the table name exists on the db
-    self.tables.filter(table_id: nil).all.each do |t|
+    self.tables.filter(table_id: nil).all.each do |user_table|
+      t = Table.new(user_table: user_table)
       t.keep_user_database_table = true
       t.destroy unless self.real_tables.map { |t| t[:relname] }.include?(t.name)
     end if dropped_tables.present? && dropped_tables.include?(nil)
@@ -1279,6 +1312,15 @@ class User < Sequel::Model
       privacy: CartoDB::Visualization::Member::PRIVACY_PUBLIC,
       exclude_shared: true,
       exclude_raster: true
+    })
+  end
+
+  # Get user owned visualizations
+  def owned_visualizations_count
+    visualization_count({
+      type: CartoDB::Visualization::Member::TYPE_DERIVED,
+      exclude_shared: true,
+      exclude_raster: false
     })
   end
 
@@ -1386,7 +1428,7 @@ class User < Sequel::Model
   end
 
   def belongs_to_organization?(organization)
-    organization_user? and self.organization_id == organization.id
+    organization_user? && organization != nil && self.organization_id == organization.id
   end
 
   def feature_flags
@@ -1482,13 +1524,33 @@ class User < Sequel::Model
       self.create_db_user
       self.grant_user_in_database
     end.join
+    self.create_own_schema
+    self.setup_schema
+  end
+
+  def create_own_schema
     self.load_cartodb_functions
     self.database_schema = self.username
     self.this.update database_schema: self.database_schema
     self.create_user_schema
     self.set_database_search_path
     self.create_public_db_user
-    self.setup_schema
+  end
+
+  def move_to_own_schema
+    self.move_to_schema(self.username)
+  end
+
+  def move_to_schema(new_schema_name)
+    if self.database_schema != new_schema_name
+      old_database_schema_name = self.database_schema
+      self.database_schema = new_schema_name
+      self.this.update database_schema: self.database_schema
+      self.create_user_schema
+      self.move_tables_to_schema(old_database_schema_name, self.database_schema)
+      self.create_public_db_user
+      self.set_database_search_path
+    end
   end
 
   def setup_schema
@@ -1499,6 +1561,15 @@ class User < Sequel::Model
     self.set_user_privileges # Set privileges
     self.set_user_as_organization_member
     self.rebuild_quota_trigger
+  end
+
+
+  def move_tables_to_schema(old_schema, new_schema)
+    self.real_tables(old_schema).each do |t|
+      self.in_database(as: :superuser) do |database|
+        database.run(%Q{ ALTER TABLE #{old_schema}.#{t[:relname]} SET SCHEMA "#{new_schema}" })
+      end
+    end
   end
 
   def reset_schema_owner
@@ -1891,7 +1962,7 @@ TRIGGER
   # Upgrade the cartodb postgresql extension
   def upgrade_cartodb_postgres_extension(statement_timeout=nil, cdb_extension_target_version=nil)
     if cdb_extension_target_version.nil?
-      cdb_extension_target_version = '0.7.2'
+      cdb_extension_target_version = '0.7.3'
     end
 
     in_database({
@@ -2222,22 +2293,82 @@ TRIGGER
     self.database_schema
   end
 
-  # return public user url -> string
-  def public_url
-    user_name = organization.nil? ? nil : username
+  # --- TODO: Extract this to a service object that handles urls
 
-    CartoDB.base_url(self.subdomain, user_name)
+  # Special url that goes to Central if active (for old dashboard only)
+  def account_url(request_protocol)
+    if CartoDB.account_host
+      request_protocol + CartoDB.account_host + CartoDB.account_path + '/' + username
+    end
+  end
+
+  # Special url that goes to Central if active
+  def plan_url(request_protocol)
+    account_url(request_protocol) + '/plan'
+  end
+
+  # Special url that goes to Central if active
+  def upgrade_url(request_protocol)
+    account_url(request_protocol) + '/upgrade'
+  end
+
+  def organization_username
+    CartoDB.subdomainless_urls? || organization.nil? ? nil : username
+  end
+
+  def organization_username
+    CartoDB.subdomainless_urls? || organization.nil? ? nil : username
   end
 
   def subdomain
-    organization.nil? ? username : organization.name
+    if CartoDB.subdomainless_urls?
+      username
+    else
+      organization.nil? ? username : organization.name
+    end
   end
+
+  # @return String public user url, which is also the base url for a given user
+  def public_url(subdomain_override=nil)
+    CartoDB.base_url(subdomain_override.nil? ? subdomain : subdomain_override, organization_username)
+  end
+
+  # ----------
 
   def name_or_username
     name.present? ? name : username
   end
 
+  def purge_redis_vizjson_cache
+    redis_keys = CartoDB::Visualization::Collection.new.fetch(user_id: self.id).map(&:redis_vizjson_key)
+    CartoDB::Visualization::Member.redis_cache.del redis_keys unless redis_keys.empty?
+  end
+
   private
+
+  # INFO: assigning to owner is necessary because of payment reasons
+  def assign_search_tweets_to_organization_owner
+    return if self.organization.nil? || self.organization.owner.nil? || self.id == self.organization.owner.id
+    self.search_tweets_dataset.each { |st|
+      st.user = self.organization.owner
+      st.save
+    }
+  rescue => e
+    Rollbar.report_message('Error assigning search tweets to org owner', 'error', { user: self.inspect, error: e.inspect })
+  end
+
+  # INFO: assigning to owner is necessary because of payment reasons
+  def assign_geocodings_to_organization_owner
+    return if self.organization.nil? || self.organization.owner.nil? || self.id == self.organization.owner.id
+    self.geocodings.each { |g|
+      g.user = self.organization.owner
+      g.data_import_id = nil
+      g.save
+    }
+  rescue => e
+    Rollbar.report_message('Error assigning geocodings to org owner, fallback to deletion', 'error', { user: self.inspect, error: e.inspect })
+    self.geocodings.each { |g| g.destroy }
+  end
 
   def name_exists_in_organizations?
     !Organization.where(name: self.username).first.nil?
@@ -2253,6 +2384,10 @@ TRIGGER
 
   def secure_digest(*args)
     Digest::SHA256.hexdigest(args.flatten.join)
+  end
+
+  def set_last_password_change_date
+    self.last_password_change_date = Time.zone.now unless new?
   end
 
 end
